+++ plugins/main/public/components/overview/compliance-table/components/subrequirements/subrequirements.tsx
@@ -34,6 +34,7 @@ import {
   TAB_VIEW_ID_EVENTS,
   TAB_VIEW_NAME_DASHBOARD,
   TAB_VIEW_NAME_EVENTS,
+  WAZUH_MODULES_ID,
 } from '../../../../../../common/constants';
 
 export class ComplianceSubrequirements extends Component {
@@ -84,16 +85,16 @@ export class ComplianceSubrequirements extends Component {
 
   getRequirementKey() {
     const mapKeys = {
-      pci: 'rule.compliance.pci_dss',
-      gdpr: 'rule.compliance.gdpr',
-      nist: 'rule.compliance.nist_800_53',
-      'nist-800-171': 'rule.compliance.nist_800_171',
-      hipaa: 'rule.compliance.hipaa',
-      tsc: 'rule.compliance.tsc',
-      'iso-27001': 'rule.compliance.iso_27001',
-      cmmc: 'rule.compliance.cmmc',
-      fedramp: 'rule.compliance.fedramp',
-      nis2: 'rule.compliance.nis2',
+      [WAZUH_MODULES_ID.CMMC]: 'rule.compliance.cmmc',
+      [WAZUH_MODULES_ID.FEDRAMP]: 'rule.compliance.fedramp',
+      [WAZUH_MODULES_ID.GDPR]: 'rule.compliance.gdpr',
+      [WAZUH_MODULES_ID.HIPAA]: 'rule.compliance.hipaa',
+      [WAZUH_MODULES_ID.ISO_27001]: 'rule.compliance.iso_27001',
+      [WAZUH_MODULES_ID.NIS2]: 'rule.compliance.nis2',
+      [WAZUH_MODULES_ID.NIST_800_53]: 'rule.compliance.nist_800_53',
+      [WAZUH_MODULES_ID.NIST_800_171]: 'rule.compliance.nist_800_171',
+      [WAZUH_MODULES_ID.PCI_DSS]: 'rule.compliance.pci_dss',
+      [WAZUH_MODULES_ID.TSC]: 'rule.compliance.tsc',
     };
     return mapKeys[this.props.section];
   }